Small fix to typecast as int

+++ ESP8266_SSD1322.cpp
@@ -879,7 +879,7 @@ void ESP8266_SSD1322::fastDrawBitmap(int16_t x, int16_t y, const uint8_t *bitmap
   // Divide by 8, as 8 pixels per byte (1 bit per pixel) unless this not, then need to add 1 extra byte
   register uint8_t wInBytes = ((w % 8) > 0) ? wDiv8 + 1 : wDiv8;
   register uint8_t wStartByte = (xDiv8 < 0 ? abs(xDiv8) : 0);
-  register uint16_t hInRows = min(SSD1322_LCDHEIGHT - y, h);
+  register uint16_t hInRows = min(SSD1322_LCDHEIGHT - y, (int)h);
   register uint16_t bytePos = 0;
   pBuf += wStartByte;  // Move start of buffer up if X < 0
 //    wInBytes -= wStartByte;
